pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "incydr"
description = "Code42's Incydr Python SDK"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Code42", email = "integrations@code42.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "boltons",
  "requests",
  "requests-toolbelt",
  "rich",
  "pydantic[dotenv]",
  "isodate",
  "python-dateutil",
]
dynamic = ["version"]

[project.optional-dependencies]
cli = ["click", "chardet"]

[project.urls]
Documentation = "https://github.com/code42/incydr_python#readme"
Issues = "https://github.com/code42/incydr_python/issues"
Source = "https://github.com/code42/incydr_python/incydr"

[project.scripts]
incydr = "incydr.__main__:incydr"

[tool.hatch.version]
path = "src/_incydr_sdk/__version__.py"

[tool.hatch.build]
sources = ["src"]
include = [
  "incydr",
  "_incydr_sdk",
  "_incydr_cli",
]

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.envs.style]
python = "3.10"  # this should always match version of "style" github action
dependencies = [
    "pre-commit",
]

[tool.hatch.envs.style.scripts]
check = "pre-commit run --all-files --show-diff-on-failure"
check-no-diff = "pre-commit run --all-files"

[tool.hatch.envs.docs]
dependencies = [
  "chardet",
  "mkdocs-autorefs",
  "mkdocs-material",
  "mkautodoc",
  "mkdocs-click",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build"
serve = "mkdocs serve"
publish = "mkdocs gh-deploy --force"
svg = "python docs/example_logging.py"


[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310","311"]
[tool.hatch.envs.test]
dependencies = [
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "pytest-httpserver",
  "pytest-lazy-fixture",
  "click",
  "chardet",
  "python-dateutil",
]
[tool.hatch.envs.test.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=incydr"
cov-report = "pytest --cov-report=xml:coverage.xml --cov-config=pyproject.toml --cov=incydr"
no-cov = "cov --no-cov"

[tool.coverage.run]
branch = true
parallel = true
omit = []

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
